split kafka catalog test for CatalogsPage.java
